Fix the number of arguments expected by the server and client demos
On my subsequent branch that adds TCP, I added a second argument that is
the protocol to use. That's why these scripts currently expect 2
arguments. However, on _this_ branch, only the first argument is used,
and that first argument is the port. I ended up b0rking this when I
extracted this branch from the rest to make PR review easier.

This commit fixes it so that the scripts expect the correct number of
arguments, which is 1 argument, the port.